This is vanilla 2.1.
The fix from #5 was reverted because it breaks the DSZ2 waterfall, making it
significantly harder to complete the stage as Sonic. (Unless you grab
the whirlwind shield and jump over it, but that's not intuitive, let
alone our place to decide. I say "was" because the revert for it vanished in #114 and I didn't
notice, so this is re-reverting it.)
The hyperwall "fix" will not be added because it will be impossible to
compete with speedruns done on 2.1.25. I don't want to force people to
use older versions of 2.1/Legacy for speedrunning because that defeats the
**entire** purpose of SRB2 Legacy. Keep in mind, Refrag did a run of 2.1
TWO MONTHS AGO that uses hyperwalls (https://youtu.be/uxHyxf0NjO4). They
will be staying.
Anything relating to movement/physics will be staying vanilla, no
exceptions.