Reapply V_HUDTRANS to the hud in st_stuff.c
Legacy's version of the HUD is quite a mess, because it uses more generic drawing macros like V_DrawScaledPatch and V_DrawPaddedTallNum instead of using the specific "INTERNAL DRAWING" macros like upstream does. When that change was made (#31), the V_HUDTRANS stopped being applied as a side effect, causing the HUD visibility setting to have no effect on that part of the HUD. This is a "temporary" fix, a more permanent fix should also be done. Remember: when in doubt, always copy whatever 2.2 does.

Also, this changes it so some of the flags are technically not coming from the right place, but that shouldn't actually matter in practice.